pyproject.toml
[tool.poetry]
name = "image-similarity"
version = "0.1.0"
description = "Find similar images."
authors = ["rodrigogiraoserrao <5621605+RodrigoGiraoSerrao@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "image_similarity"}]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
